backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
aniso8601==10.0.1
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
deprecation==2.1.0
ecdsa==0.19.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.3.0
python-keycloak==3.10.0
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==8.5.0
threadpoolctl==3.6.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
yarl==1.20.1
unstructured[docx,epub,md]~=0.16.23
chardet~=5.1.0
pandas[excel,output-formatting,performance]~=2.2.2
pandoc~=2.4
pypdfium2==4.30.0
python-docx~=1.1.0
pyyaml~=6.0.1
webvtt-py~=0.5.1